motioninsight/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
opencv-python-headless==4.11.0.86
mediapipe==0.10.33
numpy==1.26.4
Pillow==11.2.1
groq==0.22.0
python-dotenv==1.1.0
